backend/foodgram/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
Pillow==9.1.1
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1 
webcolors==1.12
python-dotenv==0.20.0
django-extra-fields==3.0.2